// src/openapi3/sdn-fabric/0.1.0/models/route.ts
// Code generated by openapi-gen. DO NOT EDIT.
/* tslint:disable */
/* eslint-disable */
import { AdditionalPropertiesUnchTarget } from './additional-properties-unch-target';

/**
 * A list of routes (single)
 */
export interface Route {

  /**
   * IP address of hop
   */
  address: string;

  /**
   * long description field
   */
  description?: string;

  /**
   * display name to use in GUI or CLI
   */
  'display-name'?: string;

  /**
   * Metric specifies the priority
   */
  metric: number;

  /**
   * subnet to match packet
   */
  prefix: string;

  /**
   * The ID of the route
   */
  'route-id': string;

  [key: string]: AdditionalPropertiesUnchTarget | number | string | undefined;
}